+++ src/helpers/escapeRegExp.ts
@@ -1,0 +1,3 @@
+export function escapeRegExp(text: string): string {
+  return text.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
+}
